+++ lithops/executors.py
@@ -49,6 +49,7 @@ from lithops.utils import FuturesList
 
 
 logger = logging.getLogger(__name__)
+CLEANER_PROCESS = None
 
 
 class FunctionExecutor:
@@ -123,9 +124,7 @@ class FunctionExecutor:
 
         self.data_cleaner = self.config['lithops'].get('data_cleaner', True)
         if self.data_cleaner and not self.is_lithops_worker:
-            spawn_cleaner = int(self.executor_id.split('-')[1]) == 0
-            atexit.register(self.clean, spawn_cleaner=spawn_cleaner,
-                            clean_cloudobjects=False)
+            atexit.register(self.clean, clean_cloudobjects=False, clean_fn=True)
 
         storage_config = extract_storage_config(self.config)
         self.internal_storage = InternalStorage(storage_config)
@@ -522,7 +521,7 @@ class FunctionExecutor:
               fs: Optional[Union[ResponseFuture, List[ResponseFuture]]] = None,
               cs: Optional[List[CloudObject]] = None,
               clean_cloudobjects: Optional[bool] = True,
-              spawn_cleaner: Optional[bool] = True,
+              clean_fn: Optional[bool] = False,
               force: Optional[bool] = False):
         """
         Deletes all the temp files from storage. These files include the function,
@@ -532,10 +531,10 @@ class FunctionExecutor:
         :param fs: List of futures to clean
         :param cs: List of cloudobjects to clean
         :param clean_cloudobjects: Delete all cloudobjects created with this executor
-        :param spawn_cleaner: Spawn cleaner background process
+        :param clan_fn: Delete cached functions in this executor
         :param force: Clean all future objects even if they have not benn completed
         """
-        os.makedirs(CLEANER_DIR, exist_ok=True)
+        global CLEANER_PROCESS
 
         def save_data_to_clean(data):
             with tempfile.NamedTemporaryFile(dir=CLEANER_DIR, delete=False) as temp:
@@ -549,6 +548,13 @@ class FunctionExecutor:
             save_data_to_clean(data)
             if not fs:
                 return
+
+        if clean_fn:
+            data = {
+                'fn_to_clean': self.executor_id,
+                'storage_config': self.internal_storage.get_storage_config()
+            }
+            save_data_to_clean(data)
 
         futures = fs or self.futures
         futures = [futures] if type(futures) != list else futures
@@ -566,8 +572,10 @@ class FunctionExecutor:
             save_data_to_clean(data)
             self.cleaned_jobs.update(jobs_to_clean)
 
+        spawn_cleaner = not(CLEANER_PROCESS and CLEANER_PROCESS.poll() is None)
         if (jobs_to_clean or cs) and spawn_cleaner:
-            sp.Popen([sys.executable, '-m', 'lithops.scripts.cleaner'], start_new_session=True)
+            cmd = [sys.executable, '-m', 'lithops.scripts.cleaner']
+            CLEANER_PROCESS = sp.Popen(cmd, start_new_session=True)
 
     def job_summary(self, cloud_objects_n: Optional[int] = 0):
         """